04282015 UT Update/optimization of prod applciations
git-svn-id: https://www.etbmanager.org/svn/pharmadex/pharmadex/trunk@324 24bb7120-ba24-2248-a177-eda5b6c43647

+++ core/src/main/java/org/msh/pharmadex/dao/iface/AdminRouteDAO.java
@@ -13,7 +13,7 @@ import java.util.List;
  * To change this template use File | Settings | File Templates.
  */
 
-public interface AdminRouteDAO extends JpaRepository<AdminRoute, Long> {
+public interface AdminRouteDAO extends JpaRepository<AdminRoute, Integer> {
 
 
 }
